+++ Assets/Scripts/StateMachine/States/GameState.cs
@@ -4,7 +4,6 @@ namespace MarkUlrich.StateMachine.States
 {
     public class GameState : State
     {
-        private bool _isSceneLoaded = false;
         private const string SCENE_NAME = "Game";
 
         public override void EnterState()
